Move type_aliases
