+++ py-polars/polars/datatypes/constants.py
@@ -22,7 +22,7 @@ from polars.datatypes import (
 
 if TYPE_CHECKING:
     from polars.datatypes.type_aliases import PolarsDataType
-    from polars.internals.type_aliases import TimeUnit
+    from polars.type_aliases import TimeUnit
 
 DTYPE_TEMPORAL_UNITS: frozenset[TimeUnit] = frozenset(["ns", "us", "ms"])
 DATETIME_DTYPES: frozenset[PolarsDataType] = frozenset(
