+++ py-polars/polars/datatypes/convert.py
@@ -67,7 +67,7 @@ else:
 
 if TYPE_CHECKING:
     from polars.datatypes.type_aliases import PolarsDataType, PythonDataType, SchemaDict
-    from polars.internals.type_aliases import TimeUnit
+    from polars.type_aliases import TimeUnit
 
     if sys.version_info >= (3, 8):
         from typing import Literal
