+++ py-polars/polars/expr/list.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
     from datetime import date, datetime, time
 
     from polars.expr.expr import Expr
-    from polars.internals.type_aliases import NullBehavior, ToStructStrategy
+    from polars.type_aliases import NullBehavior, ToStructStrategy
 
 
 class ExprListNameSpace:
