+++ py-polars/polars/series/datetime.py
@@ -11,9 +11,9 @@ if TYPE_CHECKING:
     from datetime import date, datetime, time, timedelta
 
     from polars.expr.expr import Expr
-    from polars.internals.type_aliases import EpochTimeUnit, TimeUnit
     from polars.polars import PySeries
     from polars.series.series import Series
+    from polars.type_aliases import EpochTimeUnit, TimeUnit
 
 
 @redirect(
