+++ py-polars/polars/series/string.py
@@ -8,9 +8,9 @@ from polars.series.utils import expr_dispatch
 if TYPE_CHECKING:
     from polars.datatypes import PolarsDataType, PolarsTemporalType
     from polars.expr.expr import Expr
-    from polars.internals.type_aliases import TransferEncoding
     from polars.polars import PySeries
     from polars.series.series import Series
+    from polars.type_aliases import TransferEncoding
 
 
 @expr_dispatch
