+++ py-polars/tests/unit/datatypes/test_temporal.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from zoneinfo import ZoneInfo
 
     from polars.datatypes import PolarsTemporalType
-    from polars.internals.type_aliases import TimeUnit
+    from polars.type_aliases import TimeUnit
 else:
     from polars.utils.convert import get_zoneinfo as ZoneInfo
