+++ py-polars/tests/unit/io/test_avro.py
@@ -11,7 +11,7 @@ import polars as pl
 from polars.testing import assert_frame_equal
 
 if TYPE_CHECKING:
-    from polars.internals.type_aliases import AvroCompression
+    from polars.type_aliases import AvroCompression
 
 COMPRESSIONS = ["uncompressed", "snappy", "deflate"]
