+++ py-polars/tests/unit/io/test_ipc.py
@@ -13,7 +13,7 @@ import polars as pl
 from polars.testing import assert_frame_equal, assert_frame_equal_local_categoricals
 
 if TYPE_CHECKING:
-    from polars.internals.type_aliases import IpcCompression
+    from polars.type_aliases import IpcCompression
 
 COMPRESSIONS = ["uncompressed", "lz4", "zstd"]
