+++ py-polars/tests/unit/io/test_parquet.py
@@ -18,7 +18,7 @@ import polars as pl
 from polars.testing import assert_frame_equal, assert_frame_equal_local_categoricals
 
 if TYPE_CHECKING:
-    from polars.internals.type_aliases import ParquetCompression
+    from polars.type_aliases import ParquetCompression
 
 COMPRESSIONS = [
     "lz4",
