+++ py-polars/tests/unit/operations/test_pivot.py
@@ -9,7 +9,7 @@ import polars as pl
 from polars.testing import assert_frame_equal
 
 if TYPE_CHECKING:
-    from polars.internals.type_aliases import PivotAgg
+    from polars.type_aliases import PivotAgg
 
 
 def test_pivot() -> None:
