+++ py-polars/tests/unit/operations/test_rolling.py
@@ -18,7 +18,7 @@ import polars as pl
 from polars.testing import assert_frame_equal
 
 if TYPE_CHECKING:
-    from polars.internals.type_aliases import ClosedInterval
+    from polars.type_aliases import ClosedInterval
 
 
 @pytest.fixture()
